packages/notebooks/pyproject.toml
[project]
name = "rhoai-mcp-notebooks"
version = "0.1.0"
description = "Notebooks (Workbench) component for RHOAI MCP - Kubeflow Notebook management"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Adam Miller", email = "admiller@redhat.com" }
]
maintainers = [
    { name = "Kubeflow Team", email = "kubeflow-team@redhat.com" }
]
requires-python = ">=3.10"
keywords = ["mcp", "openshift", "rhoai", "kubeflow", "notebooks", "jupyter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "rhoai-mcp-core>=0.1.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.entry-points."rhoai_mcp.plugins"]
notebooks = "rhoai_mcp_notebooks.plugin:create_plugin"

[project.urls]
Homepage = "https://github.com/admiller/rhoai-mcp-prototype"
Repository = "https://github.com/admiller/rhoai-mcp-prototype"
Issues = "https://github.com/admiller/rhoai-mcp-prototype/issues"
Changelog = "https://github.com/admiller/rhoai-mcp-prototype/blob/main/packages/notebooks/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhoai_mcp_notebooks"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["rhoai_mcp_notebooks", "rhoai_mcp_core"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.coverage.run]
source = ["src/rhoai_mcp_notebooks"]
branch = true
